feat(#98): moved SubscriptionPlan and RoleEnum to interface
