--- src/main/java/com/eshabakhov/schoodule/user/RoleEnum.java
@@ -1,39 +1,0 @@
-/*
- * © 2025-2026 Eset Shabakhov. Schoodule
- */
-package com.eshabakhov.schoodule.user;
-
-/**
- * User role enumeration.
- *
- * @since 0.0.1
- */
-public enum RoleEnum {
-
-    /** System administrator. */
-    ADMIN,
-
-    /** School director. */
-    DIRECTOR,
-
-    /** Deputy director. */
-    DEPUTY_DIRECTOR,
-
-    /** Teacher. */
-    TEACHER,
-
-    /** Student. */
-    STUDENT,
-
-    /** Basic maker. */
-    BASIC_MAKER,
-
-    /** Advanced maker. */
-    ADVANCED_MAKER,
-
-    /** Pro maker. */
-    PRO_MAKER,
-
-    /** Viewer. */
-    VIEWER
-}
